# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/transportistas:
    get:
      operationId: TransportistasController_listarActivos
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      tags: &a6
        - Transportistas
    post:
      operationId: TransportistasController_crear
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearTransportistaDto"
      responses:
        "201":
          description: ""
      tags: *a6
...
  /api/transportistas/{id}:
    put:
      operationId: TransportistasController_editar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditarTransportistaDto"
      responses:
        "200":
          description: ""
      tags: *a6
  /api/devoluciones/autorizaciones:
    post:
      operationId: AutorizacionController_crear
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearAutorizacionDto"
      responses:
        "201":
          description: ""
      tags: &a7
        - Autorizacion
    get:
      operationId: AutorizacionController_listar
      parameters:
        - name: estado
          required: false
          in: query
          schema:
            type: string
        - name: clienteId
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a7
  /api/devoluciones/autorizaciones/{id}:
    get:
      operationId: AutorizacionController_detalle
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
      tags: *a7
  /api/devoluciones/autorizaciones/{id}/reconciliacion:
    get:
      operationId: AutorizacionController_reconciliacion
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
      tags: *a7
  /api/devoluciones/autorizaciones/{id}/aprobar:
    patch:
      operationId: AutorizacionController_aprobar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
      tags: *a7
  /api/devoluciones/autorizaciones/{id}/declaracion:
    patch:
      operationId: AutorizacionController_declarar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeclararDto"
      responses:
        "200":
          description: ""
      tags: *a7
  /api/devoluciones/autorizaciones/{id}/despachar:
    patch:
      operationId: AutorizacionController_despachar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
      tags: *a7
  /api/devoluciones/autorizaciones/{id}/recibir:
    patch:
      operationId: AutorizacionController_recibir
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecibirDto"
      responses:
        "200":
          description: ""
      tags: *a7
  /api/devoluciones/autorizaciones/{id}/ingreso:
    patch:
      operationId: AutorizacionController_ingreso
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngresoDto"
      responses:
        "200":
          description: ""
      tags: *a7
  /api/devoluciones/autorizaciones/{id}/bultos/{numero}/control:
    post:
      operationId: AutorizacionController_controlar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
        - name: numero
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlarBultoDto"
      responses:
        "201":
          description: ""
      tags: *a7
  /api/devoluciones/autorizaciones/{id}/cierre:
    patch:
      operationId: AutorizacionController_cerrar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CerrarDto"
      responses:
        "200":
          description: ""
      tags: *a7
  /api/devoluciones/autorizaciones/{id}/bultos/{numero}/correccion:
    patch:
      operationId: AutorizacionController_corregir
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
        - name: numero
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorregirControlDto"
      responses:
        "200":
          description: ""
      tags: *a7
...
info:
  title: WMS Grupal — API
  description: "API del WMS Grupal. Módulo 1: Devoluciones de libros."
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CrearTransportistaDto:
      type: object
      properties: {}
    EditarTransportistaDto:
      type: object
      properties: {}
    CrearAutorizacionDto:
      type: object
      properties: {}
    DeclararDto:
      type: object
      properties: {}
    RecibirDto:
      type: object
      properties: {}
    IngresoDto:
      type: object
      properties: {}
    ControlarBultoDto:
      type: object
      properties: {}
    CerrarDto:
      type: object
      properties: {}
    CorregirControlDto:
      type: object
      properties: {}